+++ app/scripts/pages/dashboard.jsx
@@ -20,7 +20,7 @@ class Dashboard extends React.Component {
         if(mainStore.openTagManager) mainStore.toggleTagManager();
         mainStore.getTagLabels(); // Used to generate a list of tag labels
         dashboardStore.setDrawer();
-        mainStore.setRouter(this.props.router);
+        dashboardStore.setRouter(this.props.router);
     }
 
     render() {
