+++ app/scripts/stores/mainStore.js
@@ -15,7 +15,7 @@ export class MainStore {
     @observable agentKey
     @observable agentApiToken
     @observable allItemsSelected
-    @observable ancestorStatus
+    // @observable ancestorStatus
     @observable autoCompleteLoading
     @observable audit
     @observable currentUser
@@ -24,7 +24,7 @@ export class MainStore {
     @observable destination
     @observable destinationKind
     @observable device
-    @observable downloadedItems
+    // @observable downloadedItems
     @observable drawerLoading
     @observable entityObj
     @observable errorModals
@@ -66,7 +66,7 @@ export class MainStore {
     @observable projectRole
     @observable metaObjProps
     @observable responseHeaders
-    @observable router
+    // @observable router
     @observable screenSize
     @observable searchFilesList
     @observable searchFilters
@@ -76,7 +76,7 @@ export class MainStore {
     @observable searchResultsProjects
     @observable searchValue
     @observable selectedEntity
-    @observable selectedItem
+    // @observable selectedItem
     @observable showBackButton
     @observable serviceOutageNoticeModalOpen
     @observable showFilters
@@ -106,7 +106,7 @@ export class MainStore {
         this.agentKey = {};
         this.agentApiToken = {};
         this.allItemsSelected = false;
-        this.ancestorStatus = observable.map();
+        // this.ancestorStatus = observable.map();
         this.autoCompleteLoading = false;
         this.audit = {};
         this.counter = 0;
@@ -116,7 +116,7 @@ export class MainStore {
         this.device = {};
         this.destination = null;
         this.destinationKind = null;
-        this.downloadedItems = observable.map();
+        // this.downloadedItems = observable.map();
         this.drawerLoading = false;
         this.entityObj = null;
         this.errorModals = [];
@@ -160,7 +160,7 @@ export class MainStore {
         this.projectRoles = observable.map();
         this.metaObjProps = [];
         this.responseHeaders = {};
-        this.router = null;
+        // this.router = null;
         this.screenSize = {width: 0, height: 0};
         this.searchFilesList = [];
         this.searchFilters = [];
@@ -171,7 +171,7 @@ export class MainStore {
         this.searchValue = null;
         this.serviceOutageNoticeModalOpen = cookie.load('serviceOutageNoticeModalOpen');
         this.selectedEntity = null;
-        this.selectedItem = null;
+        // this.selectedItem = null;
         this.showBackButton = true;
         this.showFilters = false;
         this.showPropertyCreator = false;
@@ -201,234 +201,230 @@ export class MainStore {
     checkResponse(response) {
         return checkStatus(response, authStore);
     }
-    
-    pathFinder(kind) {
-        let kinds = {
-            'dds-project': Path.PROJECT,
-            'dds-folder': Path.FOLDER
-        }
-        return (kinds[kind])
-    }
-    
-    @action setRouter(router) {
-        this.router = router
-    }
-    
-    @action moveDownloadedItem(id, newParentId) {
-        let item = this.downloadedItems.get(id)
-        if (item && item.parentId) {
-            this.removeDownloadedItem(id, item.parentId)
-            this.addDownloadedItem(item, newParentId)
-            this.selectItem(newParentId)
-        }
-    }
-
-    @action removeDownloadedItem(id, parentId) {
-        let parent = this.downloadedItems.get(parentId)
-        let ci = parent.childrenIds.indexOf(id)
-        if(ci > -1) parent.childrenIds.splice(ci, 1)
-        let fi = parent.folderIds.indexOf(id)
-        if(fi > -1) parent.folderIds.splice(fi, 1)
-        this.downloadedItems.delete(parentId)
-        this.downloadedItems.set(parentId, parent)
-        
-        let recursiveDelete = (itemId) => {
-            let item = this.downloadedItems.get(itemId)
-            if (item) {
-                if (item.childrenIds && item.childrenIds.length > 0) {
-                    item.childrenIds.map((childId) => {
-                        recursiveDelete(childId)
-                    })
-                }
-                this.downloadedItems.delete(itemId)
-            }
-        }
-        recursiveDelete(id)
-    }
-    
-    @action addDownloadedItem(item, parentId) {
-        let parent = this.downloadedItems.get(parentId)
-        parent.open = true
-        if(parent.childrenIds) parent.childrenIds = [item.id, ...parent.childrenIds]
-        if(parent.folderIds) parent.folderIds = [item.id, ...parent.folderIds]
-        this.downloadedItems.delete(parentId)
-        this.downloadedItems.set(parentId, parent)
-        item.parentId = parentId
-        item.reload = true
-        this.downloadedItems.set(item.id, item)
-    }
-
-    @action setDownloadedItems(projects) {
-        let projectIds = []
-        projects.forEach((project) => {
-            this.downloadedItems.set(project.id, project)
-            projectIds.push(project.id)
-        })
-        this.downloadedItems.set('projectIds', projectIds);
-    }
-    
-    @action setListItems(items) {
-        this.listItems = items
-    }
-    
-    @action updateAncestorStatus(ancestors) {
-        let ancestorsDownloaded = 0
-        ancestors.forEach((ancestor) => {
-            if(this.downloadedItems.has(ancestor.id)) ancestorsDownloaded++
-        })
-        if (ancestorsDownloaded === 1) {
-            this.ancestorStatus.set('download', true)
-        } else if (ancestorsDownloaded === ancestors.length) {
-            this.ancestorStatus.set('downloadChildren', true)
-        }
-    }
-    
-    @action getAncestors(ancestors) {
-        ancestors.forEach((ancestor) => {
-            let {id, kind} = ancestor
-            if(!this.downloadedItems.has(id) && kind !== 'dds-project') {
-                this.getItem(id, this.pathFinder(kind))
-            }
-        })
-        this.ancestorStatus.set('download', false)
-    }
-    
-    @action getAncestorsChildren(ancestors) {
-        ancestors.forEach((a) => {
-            let {id, kind} = a
-            let ancestor = this.downloadedItems.get(id)
-            if(ancestor && !ancestor.childrenIds) {
-                this.getTreeListChildren(ancestor)
-            }
-        })
-        this.ancestorStatus.set('downloadComplete', true)
-        this.ancestorStatus.set('downloadChildren', false)
-    }
-  
-    @action getTreeListChildren(parent) {
-        this.loading = true;
-        let parentId = parent.id
-        let path = this.pathFinder(parent.kind)
-        let page = 1
-        this.transportLayer.getChildren(parentId, path, page)
-            .then(this.checkResponse)
-            .then((response) => {
-                const results = response.json();
-                const headers = response.headers;
-                return Promise.all([results, headers]);
-            })
-            .then((json) => {
-                let results = json[0].results;
-                let headers = json[1].map;
-                let childrenIds = []
-                let folderIds = []
-                let parentsChildren = results.map((child) => {
-                    childrenIds.push(child.id)
-                    if (child.kind === 'dds-folder') {
-                        folderIds.push(child.id)
-                        child.open = true
-                    }
-                    child.parentId = parentId
-                    if(!this.downloadedItems.has(child.id)) this.downloadedItems.set(child.id, child)
-                    return ( child );
-                });
-                let downloadedParent = this.downloadedItems.get(parentId)
-                if (downloadedParent) {
-                    downloadedParent.open = true
-                    downloadedParent.childrenIds = childrenIds
-                    downloadedParent.folderIds = folderIds
-                    this.downloadedItems.delete(parentId)
-                    this.downloadedItems.set(parentId, downloadedParent)
-                }
-                this.listItems = parentsChildren
-                this.responseHeaders = headers;
-                this.loading = false;
-            }).catch(ex =>this.handleErrors(ex))
-    }
-    
-    @action toggleTreeListItem(id) {
-        this.selectedItem = id;
-        let item = this.downloadedItems.get(id)
-        item.open = !item.open
-        this.downloadedItems.delete(id)
-        this.downloadedItems.set(id, item)
-    }
-
-    @action getItem(id, path, isSelected) {
-        mainStore.loading = true;
-        mainStore.transportLayer.getEntity(id, path)
-            .then(checkStatusAndConsistency)
-            .then(response => response.json())
-            .then((json) => {
-                if(!json.error) {
-                    let item = this.downloadedItems.get(id)
-                    if(item) {
-                        json.open = true
-                        json.childrenIds = item.childrenIds
-                        json.folderIds = item.folderIds
-                        this.downloadedItems.delete(id)
-                    }
-                    this.downloadedItems.set(id, json)
-                    if(isSelected) this.selectedItem = id
-                    mainStore.loading = false;
-                } else {
-                    json.code === 'resource_not_consistent' ? mainStore.tryAsyncAgain(getItem, [id, path, isSelected]) : mainStore.handleErrors(json);
-                }
-            }).catch(ex => mainStore.handleErrors(ex))
-    }
-
-    @action setSelectedItem() {
-        console.log('setSelectedItem');
-        let {id, path} = this.router.params
-        if (id && path) {
-            if (this.downloadedItems.has(id)) {
-                this.selectedItem = id
-            } else {
-                this.getItem(id, `${path}/`, true)
-            }
-        }
-    }
-
-    @action selectItem(itemId) {
-        console.log('selectItem');
-        let item = this.downloadedItems.get(itemId);
-        if (item) {
-            let childrenIds = item.childrenIds
-            this.selectedItem = item.id;
-            if (!childrenIds) {
-                this.getTreeListChildren(item)
-                item.reload = false
-            } else if (childrenIds.length > 0){
-                let newListItems = childrenIds.map((id) => {return(this.downloadedItems.get(id))})
-                this.listItems = newListItems
-                item.open = true
-            } else {
-                this.listItems = []
-            }
-            this.downloadedItems.delete(itemId)
-            this.downloadedItems.set(itemId, item)
-
-            if (item.kind === 'dds-project') {
-                this.project = item
-            } else {
-                this.project = this.downloadedItems.get(item.project.id)
-                mainStore.entityObj = item
-            }
-            let projPermissionsCoversion = {
-                'Project Admin': 'prjCrud',
-                'System Admin': 'prjCrud',
-                'Project Viewer': 'viewOnly',
-                'File Editor': 'flCrud',
-                'File Uploader': 'flUpload',
-                'File Downloader': 'flDownload'
-            }
-            this.projPermissions = projPermissionsCoversion[this.projectRoles.get(this.project.id)]
-
-            if (this.router) {
-                this.router.push({pathname: ('/dashboard/' + this.pathFinder(item.kind) + item.id)})
-            }
-        }
-    }
+    // 
+    // pathFinder(kind) {
+    //     let kinds = {
+    //         'dds-project': Path.PROJECT,
+    //         'dds-folder': Path.FOLDER
+    //     }
+    //     return (kinds[kind])
+    // }
+    // 
+    // @action setRouter(router) {
+    //     this.router = router
+    // }
+    // 
+    // @action moveDownloadedItem(id, newParentId) {
+    //     let item = this.downloadedItems.get(id)
+    //     if (item && item.parentId) {
+    //         this.removeDownloadedItem(id, item.parentId)
+    //         this.addDownloadedItem(item, newParentId)
+    //         this.selectItem(newParentId)
+    //     }
+    // }
+    // 
+    // @action removeDownloadedItem(id, parentId) {
+    //     let parent = this.downloadedItems.get(parentId)
+    //     let ci = parent.childrenIds.indexOf(id)
+    //     if(ci > -1) parent.childrenIds.splice(ci, 1)
+    //     let fi = parent.folderIds.indexOf(id)
+    //     if(fi > -1) parent.folderIds.splice(fi, 1)
+    //     this.downloadedItems.delete(parentId)
+    //     this.downloadedItems.set(parentId, parent)
+    //     
+    //     let recursiveDelete = (itemId) => {
+    //         let item = this.downloadedItems.get(itemId)
+    //         if (item) {
+    //             if (item.childrenIds && item.childrenIds.length > 0) {
+    //                 item.childrenIds.map((childId) => {
+    //                     recursiveDelete(childId)
+    //                 })
+    //             }
+    //             this.downloadedItems.delete(itemId)
+    //         }
+    //     }
+    //     recursiveDelete(id)
+    // }
+    // 
+    // @action addDownloadedItem(item, parentId) {
+    //     let parent = this.downloadedItems.get(parentId)
+    //     parent.open = true
+    //     if(parent.childrenIds) parent.childrenIds = [item.id, ...parent.childrenIds]
+    //     if(parent.folderIds) parent.folderIds = [item.id, ...parent.folderIds]
+    //     this.downloadedItems.delete(parentId)
+    //     this.downloadedItems.set(parentId, parent)
+    //     item.parentId = parentId
+    //     item.reload = true
+    //     this.downloadedItems.set(item.id, item)
+    // }
+    // 
+    // @action setDownloadedItems(projects) {
+    //     let projectIds = []
+    //     projects.forEach((project) => {
+    //         this.downloadedItems.set(project.id, project)
+    //         projectIds.push(project.id)
+    //     })
+    //     this.downloadedItems.set('projectIds', projectIds);
+    // }
+    // 
+    // @action updateAncestorStatus(ancestors) {
+    //     let ancestorsDownloaded = 0
+    //     ancestors.forEach((ancestor) => {
+    //         if(this.downloadedItems.has(ancestor.id)) ancestorsDownloaded++
+    //     })
+    //     if (ancestorsDownloaded === 1) {
+    //         this.ancestorStatus.set('download', true)
+    //     } else if (ancestorsDownloaded === ancestors.length) {
+    //         this.ancestorStatus.set('downloadChildren', true)
+    //     }
+    // }
+    // 
+    // @action getAncestors(ancestors) {
+    //     ancestors.forEach((ancestor) => {
+    //         let {id, kind} = ancestor
+    //         if(!this.downloadedItems.has(id) && kind !== 'dds-project') {
+    //             this.getItem(id, this.pathFinder(kind))
+    //         }
+    //     })
+    //     this.ancestorStatus.set('download', false)
+    // }
+    // 
+    // @action getAncestorsChildren(ancestors) {
+    //     ancestors.forEach((a) => {
+    //         let {id, kind} = a
+    //         let ancestor = this.downloadedItems.get(id)
+    //         if(ancestor && !ancestor.childrenIds) {
+    //             this.getTreeListChildren(ancestor)
+    //         }
+    //     })
+    //     this.ancestorStatus.set('downloadComplete', true)
+    //     this.ancestorStatus.set('downloadChildren', false)
+    // }
+    // 
+    // @action getTreeListChildren(parent) {
+    //     this.loading = true;
+    //     let parentId = parent.id
+    //     let path = this.pathFinder(parent.kind)
+    //     let page = 1
+    //     this.transportLayer.getChildren(parentId, path, page)
+    //         .then(this.checkResponse)
+    //         .then((response) => {
+    //             const results = response.json();
+    //             const headers = response.headers;
+    //             return Promise.all([results, headers]);
+    //         })
+    //         .then((json) => {
+    //             let results = json[0].results;
+    //             let headers = json[1].map;
+    //             let childrenIds = []
+    //             let folderIds = []
+    //             let parentsChildren = results.map((child) => {
+    //                 childrenIds.push(child.id)
+    //                 if (child.kind === 'dds-folder') {
+    //                     folderIds.push(child.id)
+    //                     child.open = true
+    //                 }
+    //                 child.parentId = parentId
+    //                 if(!this.downloadedItems.has(child.id)) this.downloadedItems.set(child.id, child)
+    //                 return ( child );
+    //             });
+    //             let downloadedParent = this.downloadedItems.get(parentId)
+    //             if (downloadedParent) {
+    //                 downloadedParent.open = true
+    //                 downloadedParent.childrenIds = childrenIds
+    //                 downloadedParent.folderIds = folderIds
+    //                 this.downloadedItems.delete(parentId)
+    //                 this.downloadedItems.set(parentId, downloadedParent)
+    //             }
+    //             this.listItems = parentsChildren
+    //             this.responseHeaders = headers;
+    //             this.loading = false;
+    //         }).catch(ex =>this.handleErrors(ex))
+    // }
+    // 
+    // @action toggleTreeListItem(id) {
+    //     this.selectedItem = id;
+    //     let item = this.downloadedItems.get(id)
+    //     item.open = !item.open
+    //     this.downloadedItems.delete(id)
+    //     this.downloadedItems.set(id, item)
+    // }
+    // 
+    // @action getItem(id, path, isSelected) {
+    //     mainStore.loading = true;
+    //     mainStore.transportLayer.getEntity(id, path)
+    //         .then(checkStatusAndConsistency)
+    //         .then(response => response.json())
+    //         .then((json) => {
+    //             if(!json.error) {
+    //                 let item = this.downloadedItems.get(id)
+    //                 if(item) {
+    //                     json.open = true
+    //                     json.childrenIds = item.childrenIds
+    //                     json.folderIds = item.folderIds
+    //                     this.downloadedItems.delete(id)
+    //                 }
+    //                 this.downloadedItems.set(id, json)
+    //                 if(isSelected) this.selectedItem = id
+    //                 mainStore.loading = false;
+    //             } else {
+    //                 json.code === 'resource_not_consistent' ? mainStore.tryAsyncAgain(getItem, [id, path, isSelected]) : mainStore.handleErrors(json);
+    //             }
+    //         }).catch(ex => mainStore.handleErrors(ex))
+    // }
+    // 
+    // @action setSelectedItem() {
+    //     console.log('setSelectedItem');
+    //     let {id, path} = this.router.params
+    //     if (id && path) {
+    //         if (this.downloadedItems.has(id)) {
+    //             this.selectedItem = id
+    //         } else {
+    //             this.getItem(id, `${path}/`, true)
+    //         }
+    //     }
+    // }
+    // 
+    // @action selectItem(itemId) {
+    //     console.log('selectItem');
+    //     let item = this.downloadedItems.get(itemId);
+    //     if (item) {
+    //         let childrenIds = item.childrenIds
+    //         this.selectedItem = item.id;
+    //         if (!childrenIds) {
+    //             this.getTreeListChildren(item)
+    //             item.reload = false
+    //         } else if (childrenIds.length > 0){
+    //             let newListItems = childrenIds.map((id) => {return(this.downloadedItems.get(id))})
+    //             this.listItems = newListItems
+    //             item.open = true
+    //         } else {
+    //             this.listItems = []
+    //         }
+    //         this.downloadedItems.delete(itemId)
+    //         this.downloadedItems.set(itemId, item)
+    // 
+    //         if (item.kind === 'dds-project') {
+    //             this.project = item
+    //         } else {
+    //             this.project = this.downloadedItems.get(item.project.id)
+    //             mainStore.entityObj = item
+    //         }
+    //         let projPermissionsCoversion = {
+    //             'Project Admin': 'prjCrud',
+    //             'System Admin': 'prjCrud',
+    //             'Project Viewer': 'viewOnly',
+    //             'File Editor': 'flCrud',
+    //             'File Uploader': 'flUpload',
+    //             'File Downloader': 'flDownload'
+    //         }
+    //         this.projPermissions = projPermissionsCoversion[this.projectRoles.get(this.project.id)]
+    // 
+    //         if (this.router) {
+    //             this.router.push({pathname: ('/dashboard/' + this.pathFinder(item.kind) + item.id)})
+    //         }
+    //     }
+    // }
     // 
     // @action toggleCollapseTree(router) {
     //     if (this.drawer.get('collapsed')) {
@@ -463,6 +459,10 @@ export class MainStore {
     //     this.drawer.set('width', width);
     //     this.drawer.set('contentStyle', contentStyle);
     // }
+
+    @action setListItems(items) {
+        this.listItems = items
+    }
     
     @action toggleBackButtonVisibility(bool){
         this.showBackButton = bool;
@@ -668,7 +668,7 @@ export class MainStore {
             .then((json) => {
                 this.addToast('Folder Added');
                 this.listItems = [json, ...this.listItems];
-                this.addDownloadedItem(json, id)
+                dashboardStore.addDownloadedItem(json, id)
                 this.loading = false;
             }).catch((ex) => {
             this.addToast('Failed to Add a New Folder');
@@ -706,7 +706,7 @@ export class MainStore {
 
     @action deleteItemSuccess(id, parentId, path) {
         this.loading = false;
-        this.removeDownloadedItem(id, parentId)
+        dashboardStore.removeDownloadedItem(id, parentId)
         this.listItems = this.listItems.filter(l => l.id !== id);
         this.totalItems--;
         if(this.listItems.length === 0) this.getChildren(parentId, path)
@@ -766,7 +766,7 @@ export class MainStore {
                     this.listItems.splice(index, 1, json);
                 }
                 if(this.entityObj && this.entityObj.id === id) this.entityObj = json;
-                this.downloadedItems.set(id, json);
+                dashboardStore.downloadedItems.set(id, json);
                 this.loading = false;
             }).catch((ex) => {
             this.addToast('Failed to update item');
@@ -809,7 +809,7 @@ export class MainStore {
                 } else if(!this.isListItem) {
                     this.entityObj = json;
                 }
-                this.moveDownloadedItem(id, destination)
+                dashboardStore.moveDownloadedItem(id, destination)
                 this.loading = false;
             }).catch((ex) => {
             this.addToast('Failed to move ' + type + ' to new location');
